removed unused code - start small
